RESOLVED - bug 249249: implement example that demonstrates JIRA core API https://bugs.eclipse.org/bugs/show_bug.cgi?id=249249

+++ org.eclipse.mylyn.jira.core/src/org/eclipse/mylyn/internal/jira/core/JiraClientManager.java
@@ -67,11 +67,13 @@ public class JiraClientManager {
 		if (!file.exists()) {
 			// clean up legacy data
 			File[] clients = this.cacheLocation.listFiles();
-			for (File directory : clients) {
-				File oldData = new File(directory, "server.ser"); //$NON-NLS-1$
-				if (oldData.exists()) {
-					oldData.delete();
-					directory.delete();
+			if (clients != null) {
+				for (File directory : clients) {
+					File oldData = new File(directory, "server.ser"); //$NON-NLS-1$
+					if (oldData.exists()) {
+						oldData.delete();
+						directory.delete();
+					}
 				}
 			}
 		} else {
